chore(workflow): modularize workflow code
