+++ internal/commands/workflow/workflow.go
@@ -1,22 +1,18 @@
 /*
-Copyright © 2025 NAME HERE <EMAIL ADDRESS>
+Copyright © 2025 SuprSend
 */
-package commands
+package workflow
 
 import (
 	"github.com/spf13/cobra"
 )
 
 // workflowCmd represents the workflow command
-var workflowCmd = &cobra.Command{
+var WorkflowCmd = &cobra.Command{
 	Use:   "workflow",
 	Short: "Manage workflows",
 	Long:  `Manage workflows`,
 	Run: func(cmd *cobra.Command, args []string) {
 		cmd.Help()
 	},
-}
-
-func init() {
-	rootCmd.AddCommand(workflowCmd)
 }
